splitting off tests, using nose to run
