+++ requirements.txt
@@ -3,3 +3,4 @@ Jinja2==2.7.3
 MarkupSafe==0.23
 Werkzeug==0.10.1
 itsdangerous==0.24
+nose==1.3.4
